BE hwan

+++ backend/src/main/java/swith/backend/service/PostServiceImpl.java
@@ -19,6 +19,10 @@ import swith.backend.repository.PostRepository;
 import swith.backend.repository.PostRepositorySupport;
 import swith.backend.repository.UserRepository;
 
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Optional;
+
 import static swith.backend.exception.PostExceptionType.POST_NOT_FOUND;
 
 @Service
@@ -61,6 +65,7 @@ public class PostServiceImpl implements PostService{
      * 게시글 삭제
      */
     @Override
+    @Transactional
     public void delete(Long id) {
 
         Post post = postRepository.findById(id).orElseThrow(() ->
@@ -68,11 +73,14 @@ public class PostServiceImpl implements PostService{
 
         checkAuthority(post, PostExceptionType.NOT_AUTHORITY_DELETE_POST);
 
-        for (Attachment attachment : post.getAttachments()) {
-            String uploadFileName = attachment.getUploadFileName();
-            String uploadFilePath = attachment.getUploadFilePath();
-            s3Service.deleteFile(uploadFilePath, uploadFileName);
+        if (!post.getAttachments().isEmpty()) {
+            for (Attachment attachment : post.getAttachments()) {
+                String uploadFileName = attachment.getUploadFileName();
+                String uploadFilePath = attachment.getUploadFilePath();
+                s3Service.deleteFile(uploadFilePath, uploadFileName);
+            }
         }
+
         postRepository.delete(post);
     }
 
@@ -118,5 +126,18 @@ public class PostServiceImpl implements PostService{
         Page<Post> postList = postRepositorySupport.findAllWithQuerydsl(postSearch,pageable);
 
         return postList;
+    }
+
+    @Override
+    public List<Post> getPostsByUserId(Long userId) {
+        List<Post> posts = new ArrayList<>();
+
+        Optional<List<Post>> allPosts = postRepository.findAllByUserId(userId);
+
+        if (!allPosts.isEmpty()) {
+            posts = allPosts.get();
+        }
+
+        return posts;
     }
 }
